Only merge unique indexes into the fields being validated.
We originally merged in the existing profile fields because we wanted
to prevent a user or script from removing a required field & so we
needed to know what was already set. This causes issues when we add new
validation rules for existing fields like `country` though w/o much
benefit.